Sec hailer mask ** NOT READY **
**** This only up for a code review, it isn't ready for merging yet, I'm waiting for the finished sprites, for now its using the standard gas mask as a placeholder ****

* Problem: People complain about sec silently stunning and cuffing them, sec complains that if they stop to ask someone to surrender they sanic off.

* Solution:
This commit creates a special security gasmask.
As well as functioning as a normal gas mask it has an integrated loud hailer which when activated (via attack self, verb, or action button (which means both hands are free (which means it'll actually get used unlike a hand-hailer))) emits a random compliance phrase, both as a sound effect and in the chat window.

You can download the sound effects here:
https://dl.dropboxusercontent.com/u/95696802/Commits/secmask/oggshq.rar

Compliance phrases are:
01.	HALT! HALT! HALT! HALT!
02.	Stop right there, criminal scum!
03.	Freeze, Scum Bag!
04.	Dead or alive you're coming with me.
05.	God made today for the crooks we could not catch yesterday.
06.	Don't move, Creep!
07.	Down on the floor, Creep!
08.	Prepare for justice!
09.	Running will only increase your sentence.
10.	Compliance is in your best interest
11.	Stop in the name of the Law.
// Phrases past this point require the user to tamper with the device via a screwdriver
12.	Stop or I'll bash you.
13.	Go ahead, make my day.
14.	Stop breaking the law, ass hole.
15.	You have the right to shut the fuck up.
16.	Shut up crime!
17.	Face the wrath of the golden bolt.
18.	I am, the LAW!

-Replaced the Oggs with higher quality ones, they were hard to hear in game
-Reorganised the groupings of phrases from 2 groups to three (good cop, bad cop, and shitcurity)
-Expanded the restrictor, user can now select an option to only play the good cop phrases. Also added a 'secret' way to make the mask only play shitcurity phrases

Added the finished sprites (Credit goes to Iatots for the sprites, though I did some tweaks)

Added the masks to the lockers:
-HoS has a choice between command Blue, and White
-Warden has Red
-Officers have black
-Detective, HoP, and Captain get fuck all

Moved security belts higher up the stack, having to shift through the stuff I'd end up putting in my belt all the time was just annoying.

Changed the black mask's eye covers to green, it just looked weird