Really fixed the Free Objective bug this time!
Changed get_unconvertables to search in player_list instead of mob_list - it was picking up the dead clowns on the banarium shuttle, which of course had no mind datum.
Added a check to that proc to make sure the mob has a mind datum
Added some admin log messages to notify them of errors.