A small change to how wizard rounds end when continuous wizard rounds are off:
After the game has determinded that the wizard is dead it will check if there's any traitors alive. "Traitors" in this instance means apprentices and survivors the wizard may have created with contracts/summon guns/summon magic. Thus if the round still has a chance of being interesting (survivors) and/or the wizard has a legitimate chance at revival (apprentices) the round keeps going.