Blobize edge case fix
Fixes an edge case for blobize where if you open someone's player panel and they ghost between then and when you hit the blobize button it still gives priority to them for the created blob.

Note that you can't usually get the option to blobize a player if they've left their body before you open the player panel, if you want to blobize a specific person in this way, ask for them to return to their body.