Fixes Golems not being able to use internals with their own face.
What had happened was when ABSTRACT was introduced the golem face got its own flags, and lost the flags it usually gained from inheritance that it needed to be used as internals (MASKINTERNALS and MASKCOVERSMOUTH). It now has all the flags it used to have.

Also, for no reason, it was defined twice in the exact same manner in a row.