APC overlay bugfix.
Before: If you turned off the breaker and fiddled with on/off states and turned on the breaker it wouldn't update the overlay because the check_updates wasn't doing an if(operating) check and would think it didn't need to update the overlays once you turned on the breaker.

After: Check_update() now takes into account the operating (breaker on/off) variable and accounts for it in update_overlay.

Thanks Razharas from /tg/ #coderbus for bringing this to my attention.