Updates Hair Sprites
The old sprites were simply inverted, meaning what was supposed to be the highlights showed up as shadows, and vice-versa. What I've done here is simply un-inverted it and fiddled with the levels until they were at a point that the shading variance would be visible to everyone but the blackest of black hair colors.