Rewrote how emags work on doors
Intended to allow the deconstruction and replacement of the doors without needing an RCD. Emagged doors will be unusable for at least 60 seconds, so weld+emagging still works. It's just not permanent.

Using an emag on a door will:
-Try to open the door
-Cut power for 60 seconds
-Drop bolts
-Cut bolt lights
-Effectively make the door unusable for at least a minute

Doors that have been emagged:
-Can only be opened with a crowbar or axe if the power is off and the bolts are up
-Can be hacked as normal, except the Open wire will not work
-Cannot be interfaced by silicons at all
-No longer prevent players from using tools (Can we unwelded, wires cut, etc)
-Can be deconstructed and repaired

Misc: Also made it so you can bump open doors even if the panel is open. So people can't go around opening all the panels and making the doors really annoying to open.