Changed the admin Traitor Panel command to make someone a cultist use the ticker proc - because its previous snowflake nature I wasted an hour wondering why something wasnt working.... bah
Tidied the supplypacks up for consistant names and to avoid The Bureaucracy crate etc.

Created an Innate Communication verb thats given to the cultists when converted and removed when deconverted.
It possibly works with constructs but cant really test because the make constructs procs are all kinds of messed up.
Anyway - they'll tear at their arms, smear blood on the floor, and communicate as normal per the rune. Its pretty rudimentry right now and could be improved in a few ways - see the comments.

Added a link to open the Read Tome window on the Notes page as apparently alert() is limited to a whopping three choices.

Added a built in communication system for the tomes, it works pretty much like the current communicate rune.

Fixed a load of fucked up tabbing in ritual.dm, I don't know who to blame but I'm betting it was sadbutt.