Mineral Storage Access
Mineral Storage gives you access to the ore redemption machine, instead of keeping a list of req_one_access for various jobs. The intention is to make it easier to manage accesses and make it more clear who could use the machine.

Miners and their superiors now have access to the machine. If this becomes an issue I can adjust it.

Relabeled Mineral Storeroom access to Mineral Storage since the storeroom doesn't exist any more.