Adds Radioactive Microlaser
Adds the radioactive microlaser, a device disguised as a health analyzer
used to irradiate people, as a traitor item. It costs 4 telecrystals to
purchase, and is fairly powerful in clever hands.

The strength of the radiation is determined by the 'intensity' setting,
while the delay between the scan and the irradiation kicking in is
determined by the wavelength.

Each scan will cause the microlaser to have a brief cooldown period.
Higher intensity will increase the cooldown, while higher wavelength
will decrease it.

Wavelength is also slightly increased by the intensity as well.