fixed reinforced tables
Fixes not being able to place wrenches on (unweakened) reinforced
tables. The return statement in the conditional for the wrench in
attackby was preventing control from reaching the conditional later on
that places the wrench on the table. The conditional for the wrench now
checks if it can destroy the table, and if not it continues down to
where it is then placed on the table. table_destroy is now guaranteed to
destroy the table. Also removed the now redundant check in table_destroy
and made it more portable between different kinds of tables. Also fixed
an obvious typo.