Fixes a few carp and alien sprites
What it says on the tin. It changes the regular space carp to be dark blue like they were on NT, so it fits with the megacarp. Really almost anything would have been better than the neon pink we have right now.

Fixes hunter sneak sides with weird bright borders, updates the alien pounce stances to use the latest ones.