Ports the supermatter engine from /vg/
The SM was originally developed for bay12, but I ported the /vg/ fork.
Basically, it eats heat or atom/movables and generates radiation/power.
Everything that touches it turns to dust. If it overheats, it explodes.
Typical usage involves emitters and a cooling chamber.
Currently, these are only obtainable through cargo for 100 points.
The crate comes locked and requires CE access.