Adds bolt-action rifle
Sprites from https://tgstation13.org/phpBB/viewtopic.php?f=9&t=2151

These have a bolt that needs to be worked after each shot, and they
need it to be open in order to be reloaded. There is also a stripper
clip. Their bullet does 60 brute damage, and the gun has a 5-shot
internal magazine. They cannot be worn on the back slot as there are
no back sprites provided in that file. Bayonets are not implemented
for the same reason.

The gun can be obtained as a 50% drop from ranged russian mobs (the
other 50% still drop the mateba) and from summon guns.