Reblob
A blob bursting in space will cause a random valid crewmember to be
infected.  They'll have a short time to get to a good location before
they burst.  This means a lone blob bursting in space will not end the
round unless there are no blob-valid crewmembers to choose from.