Fixes a bunch more map issues!
Fixes #7923
Fixes #7876
Fixes #7747
Fixes #7664
Fixes #7612 (#8507 can be closed to prevent any possible conflicts)
Fixes #7372 if there was ever any issue to begin with
Fixes #7304
Fixes #6563
Adds the changes from #8515 (can be closed in case it tries to cause conflicts).
Also fixes bunches of areas and makes it a bit cleaner.

SPEED MERGE FOR THE LOVE OF GOD I DON'T WANT TO HAVE TO RESTART THIS SHIT 5 TIMES.