Make objects being dragged immune to pressure differentials (space wind)
While not a solution to space wind being shit, this is just a good change that should have been done anyway to make dragging dead/dying people back onto the station from a hull breach less annoying. Also closets and other dense shit are still annoying because the wind pushes you into the closet, which pushes it backwards and out of your grasp, which is annoying and stupid, but there's no really good way to fix that right now.