Adds large cardboard boxes to the game, constructable from five sheets of cardboard.
Cardboard boxes are closets that:
*Can only carry a single person
*Are easy to destroy
*Are flammable
*Can't be welded shut

They can also can be used for more "tactical" purposes, at the same speed as walking.

Takes the opportunity to remove some object unfriendly coding.

If this is accepted they'll be a pull a little later to replace a few closets on the maps with boxes (or maybe code to have generic closets on the map to occasionally spawn as boxes instead).