Adds fudge dice
They're six sided dice with two pluses, two minusii, and two blank sides.

They're also a jokey recipe you can create with any dice and a chocolate bar. Once you eat the literal fudge die you get a fudge die no matter what kind of die you used into its creation. This is INTENTIONAL AND PART OF THE JOKE.

Adds them to the "rare dice" spawn in dice bags.

Formally defines a d6 as a separate thing from the generic parent die. the generic parent die is still a d6 for all intents and purposes except for the fact that it's now just called a generic "die" instead of a d6. This was so the recipe didn't list a d6 as needed for the fudge dice recipe when really any die would do.

Doesn't replace existent generic die with d6eses on the maps because I don't want this to conflict for a million years. I'll swap out the generic dies at some future point after this is added.